Backport: update from main
