--- .bingo/bra.mod
@@ -1,7 +1,0 @@
-module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
-
-go 1.17
-
-replace golang.org/x/sys => golang.org/x/sys v0.0.0-20220615213510-4f61da869c0c
-
-require github.com/unknwon/bra v0.0.0-20200517080246-1e3013ecaff8
