--- .bingo/cue.mod
@@ -1,5 +1,0 @@
-module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
-
-go 1.19
-
-require cuelang.org/go v0.5.0 // cmd/cue
