--- .bingo/golangci-lint.mod
@@ -1,5 +1,0 @@
-module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
-
-go 1.24.4
-
-require github.com/golangci/golangci-lint v1.64.2 // cmd/golangci-lint
