--- .bingo/jb.mod
@@ -1,5 +1,0 @@
-module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
-
-go 1.18
-
-require github.com/jsonnet-bundler/jsonnet-bundler v0.5.1 // cmd/jb
