--- .bingo/lefthook.mod
@@ -1,5 +1,0 @@
-module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
-
-go 1.20
-
-require github.com/evilmartians/lefthook v1.4.8
