+++ .bingo/variables.env
@@ -8,17 +8,5 @@ if [ -z "$GOBIN" ]; then
 fi
 
 
-BRA="${GOBIN}/bra-v0.0.0-20200517080246-1e3013ecaff8"
-
-CUE="${GOBIN}/cue-v0.5.0"
-
 DRONE="${GOBIN}/drone-v1.5.0"
-
-GOLANGCI_LINT="${GOBIN}/golangci-lint-v1.64.2"
-
-JB="${GOBIN}/jb-v0.5.1"
-
-LEFTHOOK="${GOBIN}/lefthook-v1.4.8"
-
-SWAGGER="${GOBIN}/swagger-v0.30.6-0.20240310114303-db51e79a0e37"
