+++ .citools/src/jb/go.mod
@@ -1,0 +1,20 @@
+module jb
+
+go 1.24.4
+
+tool github.com/jsonnet-bundler/jsonnet-bundler/cmd/jb
+
+require (
+	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751 // indirect
+	github.com/alecthomas/units v0.0.0-20240927000941-0f3dac36c52b // indirect
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+	github.com/fatih/color v1.18.0 // indirect
+	github.com/jsonnet-bundler/jsonnet-bundler v0.5.1 // indirect
+	github.com/mattn/go-colorable v0.1.14 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+	github.com/stretchr/testify v1.10.0 // indirect
+	golang.org/x/sys v0.33.0 // indirect
+	gopkg.in/alecthomas/kingpin.v2 v2.2.6 // indirect
+)
